Test the 5-terminal B4SOI with self-heating
This test demonstrates that the 5-terminal BSIM SOI model 4.6 works
correctly when placed in an inverter configuration.  The inverter
chosen is basically the same as our Certification_Tests/BUG_372
inverter, modified to use the BSIM-SOI 4 model instead of BSIM-SOI 3.

Self-heating is turned on. The netlist outputs the internal
temperature node to show that self heating is actually happening.

This particular test checks the 4.6.1 model.  A second test will be
created to test the 4.5.0 model the same way.

https://joseki-vm.sandia.gov/bugzilla/show_bug?id=1297
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/39