Test 5-terminal B4SOI 4.5 with self-heating
This test is the same as B4SOI/invert_b4soi.cir with two changes:  we
use the BSIM-SOI 4.5.0 instead of 4.6.1, and we have to change the
rise/fall times of the pulsed source from 5ns to 10ns.

For whatever reason, BSIM-SOI 4.5 has a convergence failure in
transient when self-heating is on and the rise time is 5ns.  4.6
didn't have that same problem.

As a regression test, though, that's OK:  this checks that the
5-terminal BSIM-SOI 4.5 works (the 4-terminal is tested by other tests
in this directory), and that one can access the temperature node via
the internal T node (using the N operator on the print line).

https://joseki-vm.sandia.gov/bugzilla/show_bug?id=1297
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/39