Modify MEASURE_CONT/STEP/FindWhenTest1DC Test
The first version of the MEASURE_CONT/STEP/FindWhenTest1DC test
had a very small compare failure on the value of LASTMEASURE for
step 0 for Windows 10.  That measure is in the test solely to make
the step-case measure tests compatible with verbose builds.  It
doesn’t actually matter what LASTMESURE measures.  So, this commit
changes it from a MAX measure to a MIN measure, which makes the
magnitude of its values closer to the other measured values.  This
should fix the test failure on Windows 10.

SON Bugs 1274 and 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313