Remove RFC_LEVEL from DERIV-WHEN, FIND-WHEN and WHEN Tests
This commit removes most of the existing subtests for the RFC_LEVEL
qualifier from the tests for TRAN mode DERIV-WHEN, FIND-WHEN and
WHEN measures.  The remaining subtests are intended to verify
that (if given on the .MEASURE line) the RFC_LEVEL value is ignored,
and that the correct warning message is emitted.

SON Bug 1314
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1314