Test FROM-TO Qualifiers for Continuous WHEN Measures
This commit adds subtests for the FROM-TO qualifiers for DC_CONT and
NOISE_CONT measure modes for WHEN measures.  It verifies that only
the rise, fall or cross values within the FROM-TO windows, starting
at the requested value (e.g., CROSS=2), are returned.  Also, for
CROSS=LAST, only the last cross within the FROM-TO window is returned.
For AC_CONT WHEN measures, it adds a subtest for the CROSS=2 case.

SON Bug 1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274