Updated gold standards for these 3 tests
These tests all attempt to take the sqrt of a negative number in the
expression library.

The old expression library would change the sign of the argument.

The new expression library uses complex arithmetic so that the answer is
imaginary.  But, as the Bsrc device just wants the real part, it returns
a zero to it.

I believe that the new expression library is doing it the right way, so
I am changing the gold standards for these tests.