Create basic regression tests for DIODE_CMC
The DIODE_CMC, while it is a formally release CMC model, comes with no
test suite of any kind.

This commit adds four simple tests of the DIODE_CMC model intended
solely to demonstrate that the model runs and produces what look like
output from a diode.

The circuit is the standard demonstration diode clipper circuit from
the Xyce users guide.  The model is a level 2002 diode with all
default parameters.  It is run in transient, dc, ac, and noise
analysis.

The noise version prints not only the inoise and onoise, but also the
device-specific DNI and DNO, and the three individual components of
one device's noise.

This is one of those "necessary-but-insufficient" test cases.  We have
to have some regression testing in the test suite, but there is
nothing against which we can compare this model's output.

Since DIODE_CMC is just an augmented version of JUNCAP200, and
JUNCAP200 was in the same testing state, I'm calling this good enough
for declaring victory on the model.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/46