Improve Tests for FIND-WHEN and WHEN for TRAN_CONT
This commit improves the testing of FIND-WHEN and WHEN measures for
TRAN_CONT measure mode.  It adds subtests (for the non-step case) for:

  1) The default case of no RISE/FALL/CROSS value on the measure line.

  2) FIND-WHEN measures.

  3) Negative values for RISE, FALL and CROSS.

  4) The use of FROM and TO qualifiers.

This set of TRAN_CONT tests now matches the test set for FIND-WHEN
and WHEN measures for AC_CONT, DC_CONT and NOISE_CONT measure modes.

SON Bugs 1274 and 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304