Test FIND-AT, FIND-WHEN and WHEN for TRAN_CONT and .STEP
This commit updates the test for the FIND-AT, FIND-WHEN and WHEN
measures for TRAN_CONT measure mode when .STEP is used.  The subtest
set is the same as the non-step case in MEASURE_CONT.  So, this
update also addresses SON Bug 1313, by testing RISE, FALL and
CROSS values <= -2.

SON Bugs 1274 and 1313
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1274
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1313