Improve Testing of When Time for WHEN Measures
This commit improves the testing of the interpolated WHEN time for
WHEN measures for AC and NOISE measure modes.  It covers the case
where the bounds for the current sweep interval bracket the
specified TO value.  This change makes the testing for DERIV-WHEN
and WHEN measures more consistent.

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304