Improve Tests for DERIV-WHEN and WHEN DC Measures
This commit improves the testing of the interpolated WHEN time for
DERIV-WHEN and WHEN measures for DC measure mode.  First, it adds
testing for the case of the TO qualifier being given, without a FROM
qualifier also being given, for a DERIV-WHEN measure.

The function for source E3 in the MEASURE_DC/DerivTest1DC.cir netlist
was changed, so that the FROM and TO qualifier tests do not return
a (valid) derivative value of -1.  The default calculation value (for
a FAILED measure) is -1 by definition.  So, it's always better to have
the successful measures not return -1, as their correct answer.

It adds additional FAILED measure tests for both DERIV-WHEN, DERIV-AT
and WHEN measures.  In particular, the case of both FROM and TO being
given, where the FROM-TO window does not overlap with the DC sweep
window is now tested.  That is a FAILED measure, and the FROM-TO values
given in stdout should be listed as the start-end DC sweep values.

Finally, it adds tests for both FROM-TO being given and valid.  In this
case, the relative orientation for FROM-TO with respect to the direction
of the DC sweep does not matter.  As an example, the FROM4 and FROM5
measures in the MEASURE_DC/DerivTest1DC.cir netlist get the same answer
at the same DC sweep value.  This is consistent with HSPICE.  The
tests in MEASURE_DC/DerivTest1DC and MEASURE_DC/STEP/DerivTest1DC
verify that is true for both ascending and descending DC sweeps.
This subtest is only done for ascending sweeps for FIND-WHEN measures
at this point.

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304