This fixes the abmpow3 test for new expression
The old expression library used a hack to avoid Nan when evaluating a
negative number raised to a fractional power.   The new expression
library does something different, which gives a slightly different (and
arguably better) answer.

The new expression library uses complex numbers under the hood.  So, a
negative number raised to a fractional power, rather than returning Nan,
returns a complex number.

The DC Bsrc only uses real numbers, however, so it just uses the real
part of the complex result.  This gives a well-behaved answer, and
avoids the old expression hack.

This commit updates the abmpow3.cir.prn.gs.pl gold-standard generator
script to be consistent with the new expression library.  It computes
the result using complex numbers and then ouputs the real part.