Improve Testing for WHEN Measures for .DC with .STEP
This commit improves the testing of WHEN measures for DC measure
mode when .STEP is used.  In particular, it verifies that the
relative orientation for FROM-TO with respect to the direction
of the DC sweep does not matter.  So, the FROM3 and FROM4 measures
get the same answer at the same DC sweep value, for a descending
DC sweep.  The test in MEASURE_DC/FindWhenTest1DC already verified
that is true for ascending DC sweeps.  This is consistent with HSPICE.

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304