Improve Testing for DERIV-WHEN and WHEN for .TRAN
This commit improves the testing of the interpolated WHEN time for
DERV-WHEN and WHEN measures for TRAN measure mode.  It covers the
case where the bounds for the current time step interval bracket the
specified FROM and/or TO values. The test accomplishes this by forcing
output at fixed time-steps, and then using -remeasure on that
previously generated .csv file.

SON Bug 1304
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1304