Test V(*) and I(*) for .PRINT AC and .PRINT NOISE
This commit updates the regression tests for SON Bug 1238 to test
that the V(*) and I(*) operators work correctly on .PRINT AC and
.PRINT NOISE lines.  In Xyce, .PRINT AC V(a) outputs two columns,
which are Re(V(A)) and Im(V(A)).  This test updates verifies that
expansion is still done correctly for the V(*) and I(*) wildcards.

SON Bug 1238
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1238