Disable expressionDoping test
The "expressionDopingExample" test uses a feature that allows
composite params to have expression-valued components.

This feature is seldom used, and does not currently work with the new
expression package.  Eric has stated that he does not expect to get
this fixed in time for the 7.2 release and that nobody really uses
this feature.  For the time being, then, the feature is being left
broken and so I'm disabling the test.  There is a second test in
another repo that needs the same treatment.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/38