Disable incorrect XDM test
This test, XDM/HSPICE/OTHER_PARSING/special_variables, incorrectly
translation of params that depend on special variables (time, temp,
vt, etc.).

Xyce's .param feature is not supposed to support this usage (it has
always been documented that way) but due to some quirks of how various
steps of expression resolution have been changed over the years, it
sorta worked for a while (mainly because the string version of the
unresolved parameter was being substituted directly into expressions
that used it).  Some holes in error reporting let that slip through.

In the new expression, this string substitution does not happen, and
so the original, documented restriction on .params is in fact
enforced.

This is therefore an invalid test and needs to be disabled.

XDM development is changing the translation of this sort of feature
and in the future versions won't be making this same error.  When that
happens, this test can be modified to test what it *should* be doing,
and then re-enabled.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/65