Test P(*1) and W(*1) Wildcards for J, M, Q and Z
This commit updates the C_T/BUG_1211_SON/multiterminal_devices test
to also test .PRINT line wildcards of the form P(*1) and W(*1) for
the J, M, Q and Z devices.  Those wildcards should output the power
for all of those multi-terminal devices.

This commit also cleans up a typo in several of the other files in
this test-subdirectory.

SON Bug 1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211