Add a script to suggest tags lists for a binary
The full Xyce test suite, even using "run_xyce_regressionMP", can take
a long time to run.  It is common during development for developers to
run abbreviated test runs with restricted tags lists that they think
capture all the tests impacted by their development.

It is also common for developers to forget to run the *entire* suite
after they are done hacking, and then find that their restricted test
set missed some tests that they broke.

And even when developers think to run the full suite, they do not
always use the same tags list that Jenkins selects, and then they wind
up missing tests.

This commit adds a script to the test suite that allows developers to
find out what tags lists Jenkins would have used to test the binary
they are running.  Jenkins runs "Xyce -capabilities" and parses out
the response, generating a set of tags that run tests for all the
capabilities reported.  In this way a developer can run a full
Jenkins-equivalent test run *before* committing and pushing.  I highly
recommend doing so, especially after long hacking runs.

In the case of limited builds (XyceNF or Xyce-opensource), Jenkins
tests not only that the build passes all tests supported by the
binary, but that it FAILS all tests that use features unsupported by
the binary.

This script is run as follows:

   suggestXyceTagList.sh [options] <binary>

where options may be "-w" to select weekly tests instead of nightly,
or "-a" to select both weekly and nightly.  The last argument should
be the path to a binary of Xyce that will be tested.

The output of the script is a set of lines of the form:

  <TAGLIST variable>=<tag list>

containing the tags appropriate for the most comprehensive testing of
that binary.

For a full XyceRad serial build, there is only one line output,
TAGLIST.

For a parallel XyceRad build you get two lines, "TAGLIST" and
"TAGLIST_KLU".  Jenkins runs the TAGLIST with the normal command line
to invoke Xyce in parallel, and TAGLIST_KLU with the additional
"-linsol klu" command line option.

For more restricted builds such as Xyce opensource, additional
"TAGLIST_xxx_SHOULDFAIL" lines are output.  If these tag lists are
used, everything run should fail.

One could conceivably run

   eval `suggestXyceTagList.sh [options] <binary>`

which will cause the shell to import all the tags lists as shell
variables that can be used when invoking run_xyce_regression, e.g.:

   eval `suggestXyceTagList.sh /path/to/Xyce`
   run_xyce_regression ... -taglist="$TAGLIST" ... /path/to/Xyce

with additional runs of run_xyce_regression to test the other tags
lists if they are not empty.

I have so far only tested this script with a serial binary, but I
think it is correct for parallel, too.  I will test it out when my
OpenMPI build completes.