Remove unnecessary use of Getopt/GetOptions
Back in July I went through and finally resolved bug 230 on joseki,
which was about tests that have Perl scripts that unnecessarily use the
"Getopt::Long" perl module and call the GetOptions function to check
whether a test script was invoked with "--verbose" or "--debug".

This usage in scripts was almost always blindly cut and pasted from some
older script.

This is an entirely useless thing to do, because there is no way to get
the top-level test harness ("run_xyce_regression") to run these
individual scripts with that option, and as a result of a change made to
the test suite to support VALGRIND way back in 2009, any test script
that did this would cause some very strange warnings to be spewed into
the stderr file when the test was run.  The presence of these
warnings (which were entirely meaningless) has confused many a developer
since it started showing up, as these warnings are never even seen until
some test fails and then a developer tries to sift through the error
files looking for why.

Most use of Getopt and GetOptions was removed in July, but I apparently
missed a couple, and then in August that couple I missed were used as a
basis for new tests.  In all cases here, not only was it useless to be
testing for "--verbose" in the script, but the script wasn't even doing
anything with the result it found --- even if it were possible to pass
such a flag to the script, it wouldn't have actually used it.

Now there are no further instances of this pattern in the test suite
from which to copy into new tests, and hopefully the whole mistake can
just be forgotten.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=230