Add tests for bug 1827 SRN
Bug 1827 SRN was all about the old expression library being unable to
resolve breakpoints for really fast pulses.  This was important for a
customer circuit, in which the rise times were 1ps.  (1.0e-12 seconds)

The logic in the old library transformed spice_pulse into a table
internally, and then used a quasi-Newton method to find breakpoints
during the calculation.  It used this method for all table sources, and
also for any source that was tranformed into a table.

That algorithm was a fine idea for conditional statements, but it made
no sense for purely deterministic sources.  Purely determinstic sources
(such as SPICE pulse and table) have all the information they need to
determine ALL their breakpoints in the user-specification of the source.
They don't need to compute them via a numerical method.

Also, even if this algorithm had been approprite,  it was unfortunately
written poorly, and had at least one magic number in it that made
resolution of a 1ps rise time impossible.

The new expression library avoids all of this.  The spice_pulse function
inside of it produces breakpoints using the parameters of the pulse, so
as long as they specify a valid pulse, the breakpoints will be
correct.  The only tolerance at play is the breakpoint tolerance used by
the time integrator.

The new expression library TABLE function has also been improved.  So,
even if spice_pulse in the new library were internally turned into a
table, it would be able to resolve 1ps (and smaller) separated entries.
Like the spice_pulse, the TABLE in the new library uses the entries of
the table to determine breakpoints, rather than trying to determine them
via a poorly conceived numerical calculation.  Similar to spice_pulse,
the only relevant tolerance is the breakpoint tolerance used by the
time integrator.

Bug 1827 didn't have a circuit attached to it, and the circuit that
prompted the bug would have been much larger than necessary to certify
this bug.    So, based on the comments in the bug, I created two very
simple tests, one for spice_pulse and one for table.

Both test circuits consist of a Bsrc and a resistor.  And, both specify
expression-defined sources in the Bsrc that produce breakpoints that are
less than 1ps apart.    They produce them correctly.