Certification tests for bugs 1173 and 1176
These are simple RC filter circuits, with a noise source specified by a
giant table file. (100,000 points).

The bug1173son.cir netlist uses the table file with the PWL Vsrc.

The bug1176son.cir netlist is the same circuit, except that the table
file is used in a Bsrc via the expression library.

Note, bug 1173 is a closed bug, and was closed in 2019.  For that bug I
completely rewrote Xyce's breakpoint handling to be more efficient, and
I chose to have the existing test suite, which already used breakpoints
extensitvely, verify it.    But  in retrospect
I really should have created at least one test like this.

The bug 1176 test is completely new.  Bug 1176 is not closed but will be
soon.  The closure of bug 1176 is enabled by the new expression library.

Both tests are pretty fast, despite the giant table source.  On my laptop,
in serial, the bug 1173 test(Vsrc) runs in about 4 or 5 seconds and the
bug 1176 test(Bsrc) runs in about 3 seconds.

I use the ".options output initial_interval" command to keep the output
file small.  Otherwise it would have 0.5M data points in it, and
xyce_verify would be really slow.  Instead the output files have 300
points each, which xyce_verify can handle quickly.