Test ? Wildcards on .PRINT Lines
This commit adds tests for the use of the ? wildcard on .PRINT DC
lines.  That wildcard means "any single character".

In addition, the tests for the use of ? in node names and device
names, and in expressions, needed to be updated.  Examples are
node 1? and device R1?.  Those usages are not recommended, but Xyce
should do something sensible in those cases.  The Xyce behavior
is now that, for example, .PRINT DC V(1?) is a request for all
two-character nodal voltages that start with '1'.  It is not an
explicit request for the voltage at node 1?, if that node exists in
the netlist. The same now applies, for example, to .PRINT DC I(R1?)
P(R1?) W(R1?).  R1? is now treated as a wildcard request for all
three-character device names that start with "R1".  Finally, the ?
character is not considered to be a wildcard if, for example, V(1?)
is used within an expression on a .PRINT line

SON Bugs 1211 and 1319
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1211
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1319