Tests for AGAUSS and AUNIF with sampling
Issue #84

This is the first set of tests for using expression-based, Hspice-style
random operators in combination with .SAMPLING.

In order to keep things more easily organized, I chose to create a new
SAMPLING directory, called SAMPLING2.  After doing that, to create these
tests, I copied over as many as I could from SAMPLING and then modified
to use the new specification.  This necessarily excluded the test that
for gamma distributions, as I don't have an operator for that
distribution. (I'm not sure that Hspice ever bothered with that).

Typical changes to these tests are small.  For example, the
"simpleNormal.cir" test had these lines change from this:

R1 1 2 1K

.SAMPLING
+ param=R1
+ type=normal
+ means=3K
+ std_deviations=1K

to this:

R1 1 2 {agauss(3K,1K,1)}

.SAMPLING
+ useExpr=true

To connect the random operators to .SAMPLING, the .SAMPLING parmaeter
"useExpr" must be set to true.  By default it is false.  When it is
false, the .SAMPLING code will assume that there must be a list of
comma-separated params set on the .SAMPLING line.  I may change that
later, but for now that is how it is.  The code can only deal with one
specification or the other.  It cannot currently handle both.

Another small comment.  The "-hspice-ext all" command line option currently will
toggle between having ramdom expression operators initialize to a random
value, or initialize to the mean.    ie, if "-hspice-ext all" is
specified then {agauss(3k,1k,1)} will result in this expression being
initialized to 3k.  If it is not present, then it is initialzed to a
random number from that distribution.

This -hspice-ext all option is *mostly* irrelevant when .SAMPLING with
useExpr=true is present in the netlist, because the values from SAMPLING
will override the initial value before any calculations happen.
HOWEVER, there is one small complication.  That is that if hspice-ext
all is NOT present, then the random seed will be initialized 2x, and
this will result in two random seeds being printed to standard output.
For most use cases, this is not very important.  But, if you want to
re-run the same sample study with the same seed, you have to be sure to
use the correct one (the second one).

This issue had the potential to mess up the "seed" based tests in this
directory, where the netlist is run 2x and then diff'd.  So, for the
"seed" based tests, the shell script had to be modified to use
"-hspice-ext all".  For all the others, it doesn't matter.