Fix malloc Errors in MIXED_SIGNAL VPI Test
This commit fixes careless malloc usage errors in the .c file for
the MIXED_SIGNAL/VPI/runXyceWithDAC test.  For example, the call to
xyce_getNumDevices() call returns the maximum length of all device
names of the requested type.  That max length is then used in a
malloc() call to allocate arrays that hold C-style (null-terminated)
strings copied from the C++ std::string.  If the max size of all
strings is N, then the code must allocate N+1 bytes in the C-style
array to assure that there's room for the null termination.
Otherwise, the strings that are exactly that max length will
overrun the array.  This was causing valgrind errrors on RHEL7.

SON Bugs 1052 and 1062
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1052
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1062