Add .embeddedsampling tests for random operators
Issue #84

This is very similar to the tests I recently added for .SAMPLING, except
that this set of tests is for embedded sampling (simultaneous
propagation) rather than conventional sampling.  As before, I created a
new directory, EMBEDDED2, into which I copied and modified relevant
tests from the EMBEDDED directory.  This had to exclude  things like
simpleGamma and the truncated distribution tests, which don't have an
analog in the expression ops.

In the EMBEDDED directory, there are several exlcuded tests.  These are
tests that I had copied from SAMPLING, but then could not run with
embedded sampling due to unsupported analysis.  So, for example, AC, HB
and sensitivity analysis all currently do not work with embedded
sampling.  AC and HB require block solves, and so does embedded
sampling, so this means that you wind up with hierarchical block
systems.  I didn't have time to figure out how to do that.  It may not
be possible with the current linear algebra block objects.

Anyway, for those excluded tests, instead of copying them from EMBEDDED
To EMBEDDED2, only to then exclude them again, I didn't bother.  I think
I left the ones in EMBEDDED in place as a reminder to myself, and I
don't need two reminders.

ALso, when doing this, I found at least 2 potential bugs that I'll have
to dig into tomorrow.  Both of these bugs were revealed by the
embedded_cmosInverter.cir test.

(1) Initially, the embedded_cmosInverter test failed badly.  It has
two uncertain parameters (vth0 in both nmos and pmos model cards).
For the nmos:vth0 parameter, it has a uniform random distribution from
+0.4 to +0.7.  The pmos:vth0 parameter has the mirror image uniform
random distribution from -0.4 to -0.7.  I initially set up these
parameters as:

nmos:vth0={unif(0.55,0.15)}
pmos:vth0={-unif(0.55,0.15)}

It turns out that the unary minus in the vth0 parameter didn't work
correcetly; it was producing a positive set of values for pmos:vth0.
So, that is bug 1.

(2) To work around bug 1, I changed pmos:vth0 to:
pmos:vth0={unif(-0.55,-0.15)}

When I did that, the values for pmos:vth0 were correctly negative.  But
it still failed the xyce_verify stage.  The gold standard (which just
compares means, std dev, etc, not the individual samples) was generated
with the old specification.  It was consistently 2-3% off in the
comparison.  So, the two specifications are not quite the same.

I am sure I can figure both of these out, but it won't happen today.