Temporarily removing these from parallel testing
In parallel, there is a problem with using expression-based
random operators with analysis types such as SAMPLING and
EMBEDDEDSAMPLING.  It only applies to device parameters.
Global parameters are OK.

It will probably take me more than 1 test cycle to fix this, so
for now I am disabling these tests in parallel.