Add "transline" tag to all TRA circuits
There are a number of circuits in the test suite that use the lossless
transmission line but are not tagged as such.

As I am working on some code that impacts the transmission line
efficiency, I want to be able to run all the tests repeatedly to see the
impact I have.  This change enables that.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/83