Add another embedded sampling test
This test is the same as a similar test in the EMBEDDED directory,
except that it uses the random expression operators instead of the
"param=" way of specifying random parameters.

It also is the first test in which one of the random device
parameters is equal to an arithmetic expression that happens to include
a random operator (aunif), rather than simply being equal to a random
operator.    Specifically, in the pmos model card, the vth0 parameter is
set as follows:

vth0 ={-aunif(0.55,0.15)}

In other words, it is -1*aunif, rather than just aunif.

Before my code changes today, this would have failed, because the
DeviceMgr::setParameterRandomExpressionTerms function would have set
vth0 equal to aunif rather than just setting the internal operator to
the value of aunif.

Anyway, this passes now, so I can un-exclude it.