Wait if needed after file write
On HPC with distributed IO this script may try to run
a system command on a file that hasn't finished writing
to disk.  Address this by checking that the just written
files are ready.  If they are not then the script waits
for 2 seconds then continues.  If the files are ready
then the script immediately tries to run diff as before.