Add offet for outputs with zero crossings
In these two circuits a Xyce calcualted derivative and
a finite difference calculated derivative are part of the output
The tests would fail on HPC where the derivative just goes from
zero to non-zero, and when the difference between the two
values crossed zero.  I experimented to find the minimal
offset that would let these tests pass so that I didn't
mask an underlying error.  When the results are plotted one
can see that a single zero crossing where the outputs
differ by about 4e-6 contributes to the majority of
the relative error.