Tests for bug 794 SON
These are taken directly from the bug report.  test1 was submitted by
Pete, and uses a G-source.  test2 was submitted by Tom and uses a
B-source.  They both use the same underlying code in the new expression
library, and they implement identical circuits, so with respect to this
bug, if one passes then the other should as well.