Add test to support exp=constant=exp(1.0) in expr
I had neglected to include this feature in the new expression library.
In the old library, the keyword "exp" by itself, without any
parens or arguments following it, should be treated as a constant =
exp(1.0).

Part of the reason that I did not implement it before now in the new
library is that there has not been any regression test using this
capability.   I only discovered it by reading the user guide.

This commit is intended to fix that gap in testing.  The circuit is
ultra simple, with a resistor set via an expression to be dependent on
the constant "exp".