Add new expression-based sampling test, multi-ops
The point of this test is to ensure that when multiple random operators
are inside of an arithmetic expression, it will treat each operator
separately.

I chose to test this by creating two equivalent netlists.  One of them
uses global parameters and sets up two unique AGAUSS operators, one for
each param.  Then these params are used in the expression that sets the
value of R1:R.

The other netlist doesn't use global params at all.  It just puts the
two agauss operators directly into the expression for R1:R.

These two approaches should result in an identical (nearly) AST tree for
R1:R.  But the machinery for setting the values of the AGAUSS operators
is different.  So the test ensures that the two paths thru the Xyce
source code do the same thing.