Tests for FREQ in a device expression
When working on documentation today, I read over the ABM chapter in the
user's guide.

It stated that using "FREQ" in an ABM device was not allowed.  I wasn't
sure if that was true or not.  Based on my recent expression work,
it seemed as though it should work.

I ran some experiments, and FREQ does appear to work.  So, I have
created two tests to demonstrate the capability.  One (RC_simple) has an
expression-valued resistor, where the expression depends on FREQ.

The other (abmfreq1) has a Bsrc that uses FREQ in its expression.

They both work, but the Bsrc version required that I check for
divide-by-zero errors, for when FREQ=0.0.  This was the case during the
DCOP calculation, which initially caused abmfreq1 to fail the DCOP
phase.