Add expression sampling test w/FOUR random params
Issue #84

Earlier this week I fixed the code for expression-based sampling to work
in parallel.  Or, so I thought, as it appeared so with the tests that
were present in the test suite.    However, today I attempted to run a
parallel case with a larger number of parameters and it crashed badly.

I managed to fix that problem in an earlier commit.  Now, I am adding a
test that proves it.  This test is a simple voltage divider, but with
four uncertain resistors in series.  This test did not run properly
prior to today's commit.