Fix Careless Typo in MEASURE_DC/ErrorFuncTest1DC Test
This commit corrects an error in the MEASURE_DC/ErrorFuncTest1DC
test.  Two of the measures had duplicate measure names, which
produced a warning message that I had previously ignored.  Those
two measures (which test FROM and TO) now have different names,
and the appropriate "gold files" are now actually correct.  This
"oops" did not appear in any of the other MEASURE tests for AC, DC,
NOISE or TRAN measures.  In  particular, the corresponding test in
MEASURE_DC/STEP was correct and would have caught issues with the
ERR2 measure.

SON Bug 1279
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1279