Bsim3 test for .STEP thru .global_param and .param
Issue #1

On SON bugzilla:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

With the new expression library, .params are not really constant
anymore.  They still cannot be changed directly via .STEP.  But, they
can be changed indirectly if they depend on a .global_param and the
.STEP sweep applies to the global.  In other words this works:

.global_param n1l=0.35u
.param n1lpar={n1l}
.step lin N1L 0.35u 0.4u 0.01u
.model n1 NMOS L={N1LPAR}

but this will not work at present:

.param n1lpar=0.35u
.step lin n1lpar 0.35u 0.4u 0.01u
.model n1 NMOS L={N1LPAR}

The second case will not work b/c there isn't enough infrastructure in
Xyce to track .param objects.  But the first case works, and that is the
case that is tested by this new test.