More tests for bug 1190 SON
Issue #1

This is about making .param and .global_param more similar.  They are
not identical yet and might not be; we will have to see about that.  But
one important improvement is that with the new expression library,
.params are no longer constants and can depend on global_params that can
change.

I previously committed a single test for this issue, for the bsim3.
With this commit there is now 4 simple tests.  Two for mutual inductors, one
for the bsim3 (that I previously committed) and one for the bsim4.

All of them perform .STEP sweeps.  The "test" circuits all do the .STEP
sweeps over global parameters.  These global_params are used on the RHS
of .param statements, and the .params are actually used to set the model
parameter values.  The "baseline" circuits also do .STEP sweeps, but
don't use .params in any way, at least with respect to the swept
variable.