Test for bug 1085 SON
This issue was fixed by the new expression library.  The specific title
of the issue is:

"Expression library mishandles ".FUNC" functions with names
that begin with I and have two letters"

I somehow didn't mark this one as being dependent on bug 1249 (the
new expression library issue), and had marked the target milestone as
"indeterminate".   So, as a result I overlooked that this probably got
fixed for release 7.2, and I didn't do any work to close it for 7.2.
Oh well.  It is fixed.

But, better late the never.  This test proves that it is fixed.  It has
two circuits.  The first one was attached to the original bug, and used
to fail because of a .func named "I0".  The second is an identical
circuit, except that the offending .FUNC has been renamed to "test",
which the old expression library would have handled just fine.