Test for bug 50 SON  (HERTZ synonym for FREQ)
In addition to bug 50 on the SON, this is also being tracked by
Issue #93 in gitlab.

This commit adds a few more tests to the ABM_FREQ directory.  This is a
relatively recent set of tests, which proved that FREQ worked fine in
expressions, even in Bsrcs

This commit copies the 2 FREQ-based tests, and makes them HERTZ-based
tests.  Otherwise the 2 new tests are identical to the old 2 tests.