Test HERTZ and TEMPER on .PRINT Outside of Expressions
This commit tests the use of TEMPER and HERTZ (which are now
synonyms for the TEMP and FREQ special variables, respectively)
on .PRINT lines when those two special variables are outside of
an expression.  SON Bug 1212 already added tests for the FREQ
special variable on .PRINT lines.

SON Bugs 50 and 1212
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=50
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1212