Make suggestXyceTagList exclude admssens consistently
suggestXyceTagList figures out what tags to use based on the output of
"Xyce -capabilities", which in turn is based on preprocessor symbols
defined by the configuration process.

"--disable-admsmodels" and "--disable-adms_sensitivities" are
independently selectable, but the latter does absolutely nothing if
the former is already given.

"Xyce -capabilities" outputs that ADMS sensitivities are compiled in
even when ADMS devices don't exist.  I'm going to fix that, too, but
for now I'm making suggestXyceTagList not even look for it if it knows
ADMS devices aren't there.  Thus "-adms" always implies
"-adms-admssens", but "?adms" could have either "?admssens" or
"-admssens" depending on how Xyce is compiled.